refactor: reorganize job namespaces and increase Horizon job settings
- Move job classes into categorized namespaces (`Recipe` and `Menu`) for better structure.
- Update Horizon configuration to increase memory limit (512MB to 1024MB) and job timeout (300s to 900s).
