+++ app/Jobs/Recipe/ImportRecipeJob.php
@@ -2,7 +2,7 @@
 
 /** @noinspection DuplicatedCode */
 
-namespace App\Jobs;
+namespace App\Jobs\Recipe;
 
 use App\Enums\QueueEnum;
 use App\Http\Clients\HelloFresh\Responses\RecipesResponse;
