+++ config/horizon.php
@@ -245,9 +245,9 @@ return [
             'maxProcesses' => 4,
             'maxTime' => 0,
             'maxJobs' => 0,
-            'memory' => 512,
+            'memory' => 1024,
             'tries' => 1,
-            'timeout' => 300,
+            'timeout' => 900,
             'nice' => 0,
         ],
     ],
